+++ distorch_2.py
@@ -16,6 +16,8 @@ import inspect
 from collections import defaultdict
 import comfy.model_management as mm
 import comfy.model_patcher
+from .device_utils import get_device_list, soft_empty_cache_multigpu
+from .model_management_mgpu import multigpu_memory_log, track_modelpatcher
 
 safetensor_allocation_store = {}
 safetensor_settings_store = {}
@@ -59,7 +61,6 @@ def register_patched_safetensor_modelpatcher():
 
         def new_partially_load(self, device_to, extra_memory=0, full_load=False, force_patch_weights=False, **kwargs):
             """Override to use our static device assignments"""
-            from .device_utils import multigpu_memory_log, track_modelpatcher
             global safetensor_allocation_store
 
             debug_hash = create_safetensor_model_hash(self, "partial_load")
@@ -181,7 +182,6 @@ def analyze_safetensor_loading(model_patcher, allocations_string):
     Analyze and distribute safetensor model blocks across devices
     Target for refactor back into one function once stability for CLIP is established.
     """
-    from .device_utils import get_device_list
     DEVICE_RATIOS_DISTORCH = {}
     device_table = {}
     distorch_alloc = allocations_string
@@ -389,7 +389,6 @@ def analyze_safetensor_loading_clip(model_patcher, allocations_string):
     All other logic and UX (logging, etc.) is identical to the original.
     Target for refactor once stability for CLIP is established.
     """
-    from .device_utils import get_device_list
     DEVICE_RATIOS_DISTORCH = {}
     device_table = {}
     distorch_alloc = allocations_string
@@ -805,7 +804,6 @@ def override_class_with_distorch_safetensor_v2(cls):
     class NodeOverrideDisTorchSafetensorV2(cls):
         @classmethod
         def INPUT_TYPES(s):
-            from .device_utils import get_device_list
             inputs = copy.deepcopy(cls.INPUT_TYPES())
             devices = get_device_list()
             compute_device = devices[1] if len(devices) > 1 else devices[0]
@@ -902,7 +900,6 @@ def override_class_with_distorch_safetensor_v2_clip(cls):
     class NodeOverrideDisTorchSafetensorV2Clip(cls):
         @classmethod
         def INPUT_TYPES(s):
-            from .device_utils import get_device_list
             inputs = copy.deepcopy(cls.INPUT_TYPES())
             devices = get_device_list()
             default_device = devices[1] if len(devices) > 1 else devices[0]
@@ -1000,7 +997,6 @@ def override_class_with_distorch_safetensor_v2_clip_no_device(cls):
     class NodeOverrideDisTorchSafetensorV2ClipNoDevice(cls):
         @classmethod
         def INPUT_TYPES(s):
-            from .device_utils import get_device_list
             inputs = copy.deepcopy(cls.INPUT_TYPES())
             devices = get_device_list()
             default_device = devices[1] if len(devices) > 1 else devices[0]
