+++ model_management_mgpu.py
@@ -1,0 +1,333 @@
+"""
+Model Management Extensions for MultiGPU
+Extends ComfyUI's model management with multi-device capabilities and lifecycle tracking.
+"""
+
+import torch
+import logging
+import hashlib
+import psutil
+import comfy.model_management as mm
+import gc
+from datetime import datetime, timezone
+import server
+import weakref
+import platform
+import ctypes
+import comfy.model_patcher
+from collections import defaultdict
+
+logger = logging.getLogger("MultiGPU")
+
+# ==========================================================================================
+# Model Analysis and Store Management (DisTorch V1 & V2)
+# ==========================================================================================
+
+# DisTorch V2 SafeTensor stores
+safetensor_allocation_store = {}
+safetensor_settings_store = {}
+
+# DisTorch V1 GGUF stores (backwards compatibility)
+model_allocation_store = {}
+
+def create_safetensor_model_hash(model, caller):
+    """Create a unique hash for a safetensor model to track allocations"""
+    if hasattr(model, 'model'):
+        actual_model = model.model
+        model_type = type(actual_model).__name__
+        model_size = model.model_size() if hasattr(model, 'model_size') else sum(p.numel() * p.element_size() for p in actual_model.parameters())
+        first_layers = str(list(model.model_state_dict().keys() if hasattr(model, 'model_state_dict') else actual_model.state_dict().keys())[:3])
+    else:
+        model_type = type(model).__name__
+        model_size = sum(p.numel() * p.element_size() for p in model.parameters())
+        first_layers = str(list(model.state_dict().keys())[:3])
+    
+    identifier = f"{model_type}_{model_size}_{first_layers}"
+    final_hash = hashlib.sha256(identifier.encode()).hexdigest()
+    logger.debug(f"[MultiGPU DisTorch V2] Created hash for {caller}: {final_hash[:8]}...")
+    return final_hash
+
+def create_model_hash(model, caller):
+    """Create a unique hash for a GGUF model to track allocations (DisTorch V1)"""
+    model_type = type(model.model).__name__
+    model_size = model.model_size()
+    first_layers = str(list(model.model_state_dict().keys())[:3])
+    identifier = f"{model_type}_{model_size}_{first_layers}"
+    final_hash = hashlib.sha256(identifier.encode()).hexdigest()
+    logger.debug(f"[MultiGPU_DisTorch_HASH] Created hash for {caller}: {final_hash[:8]}...")
+    return final_hash
+
+def prune_distorch_stores():
+    """Prune stale allocation/settings entries not tied to active models."""
+    multigpu_memory_log("distorch_prune", "start")
+    active_hashes_v2 = set()
+    active_hashes_v1 = set()
+    
+    for lm in mm.current_loaded_models:
+        mp = lm.model
+        if mp is not None:
+            active_hashes_v2.add(create_safetensor_model_hash(mp, "prune_check_v2"))
+            active_hashes_v1.add(create_model_hash(mp, "prune_check_v1"))
+
+    # V1 pruning
+    stale_v1 = set(model_allocation_store.keys()) - active_hashes_v1
+    if stale_v1:
+        logger.info(f"[MultiGPU_Memory_Management] Pruning {len(stale_v1)} DisTorch V1 entries")
+        for k in stale_v1:
+            del model_allocation_store[k]
+
+    # V2 pruning
+    for store, name in ((safetensor_allocation_store, "allocation"), (safetensor_settings_store, "settings")):
+        stale_v2 = set(store.keys()) - active_hashes_v2
+        if stale_v2:
+            logger.info(f"[MultiGPU_Memory_Management] Pruning {len(stale_v2)} V2 {name} entries")
+            for k in stale_v2:
+                del store[k]
+    
+    multigpu_memory_log("distorch_prune", "end")
+
+# ==========================================================================================
+# Memory Logging Infrastructure
+# ==========================================================================================
+
+_MEM_SNAPSHOT_LAST = {}
+_MEM_SNAPSHOT_SERIES = {}
+
+def _capture_memory_snapshot():
+    """Capture memory snapshot for CPU and all devices"""
+    # Import here to avoid circular dependency
+    from .device_utils import get_device_list
+    
+    snapshot = {}
+    
+    # CPU
+    vm = psutil.virtual_memory()
+    snapshot["cpu"] = (vm.used, vm.total)
+    
+    # GPU devices
+    devices = [d for d in get_device_list() if d != "cpu"]
+    for dev_str in devices:
+        device = torch.device(dev_str)
+        total = mm.get_total_memory(device)
+        free_info = mm.get_free_memory(device, torch_free_too=True)
+        system_free = free_info[0] if isinstance(free_info, tuple) else free_info
+        used = max(0, total - system_free)
+        snapshot[dev_str] = (used, total)
+
+    return snapshot
+
+def multigpu_memory_log(identifier, tag):
+    """Record timestamped memory snapshot with delta logging"""
+    if identifier == "print_summary":
+        for id_key in sorted(_MEM_SNAPSHOT_SERIES.keys()):
+            series = _MEM_SNAPSHOT_SERIES[id_key]
+            logger.mgpu_mm_log(f"=== memory summary: {id_key} ===")
+            for ts, tag_name, snap in series:
+                parts = []
+                cpu_used, cpu_total = snap.get("cpu", (0, 0))
+                parts.append(f"cpu={cpu_used/(1024**3):.2f}/{cpu_total/(1024**3):.2f}")
+                for dev in sorted([k for k in snap.keys() if k != "cpu"]):
+                    used, total = snap[dev]
+                    parts.append(f"{dev}={used/(1024**3):.2f}/{total/(1024**3):.2f}")
+                ts_str = ts.strftime("%Y-%m-%dT%H:%M:%S.%f")[:-3] + "Z"
+                logger.mgpu_mm_log(f"{ts_str} {id_key} {tag_name} | " + " | ".join(parts))
+        return
+
+    ts = datetime.now(timezone.utc)
+    curr = _capture_memory_snapshot()
+    
+    # Store in series
+    if identifier not in _MEM_SNAPSHOT_SERIES:
+        _MEM_SNAPSHOT_SERIES[identifier] = []
+    _MEM_SNAPSHOT_SERIES[identifier].append((ts, tag, curr))
+    
+    # Compute delta
+    if identifier in _MEM_SNAPSHOT_LAST:
+        prev_tag, prev = _MEM_SNAPSHOT_LAST[identifier]
+        keys = set(prev.keys()) | set(curr.keys())
+        ordered = ["cpu"] + sorted([k for k in keys if k != "cpu"])
+        parts = []
+        for k in ordered:
+            p_used, _ = prev.get(k, (0, 0))
+            c_used, _ = curr.get(k, (0, 0))
+            delta = c_used - p_used
+            sign = "+" if delta >= 0 else "-"
+            parts.append(f"{k}={sign}{abs(delta)/(1024**3):.2f}")
+        logger.mgpu_mm_log(f"{identifier} {tag} - {prev_tag}: " + " | ".join(parts))
+    else:
+        # Baseline
+        ordered = ["cpu"] + sorted([k for k in curr.keys() if k != "cpu"])
+        parts = []
+        for k in ordered:
+            c_used, _ = curr.get(k, (0, 0))
+            parts.append(f"{k}=+{c_used/(1024**3):.2f}")
+        logger.mgpu_mm_log(f"{identifier} {tag} - <baseline>: " + " | ".join(parts))
+
+    _MEM_SNAPSHOT_LAST[identifier] = (tag, curr)
+
+def clear_memory_snapshot_history():
+    """Clear stored memory snapshot history"""
+    multigpu_memory_log("mem_mgmt", "pre-history-clear")
+    _MEM_SNAPSHOT_LAST.clear()
+    _MEM_SNAPSHOT_SERIES.clear()
+    logger.debug("[MultiGPU_Memory_Management] Memory snapshot history cleared")
+    multigpu_memory_log("mem_mgmt", "post-history-clear")
+
+# ==========================================================================================
+# ModelPatcher Lifecycle Tracking
+# ==========================================================================================
+
+_MGPU_TRACKED_MODELPATCHERS = weakref.WeakSet()
+
+def track_modelpatcher(model_patcher):
+    """Register ModelPatcher for lifecycle tracking"""
+    if isinstance(model_patcher, comfy.model_patcher.ModelPatcher):
+        if model_patcher not in _MGPU_TRACKED_MODELPATCHERS:
+            _MGPU_TRACKED_MODELPATCHERS.add(model_patcher)
+            logger.debug(f"[MultiGPU_Lifecycle] Tracking ModelPatcher {id(model_patcher)} (tracked={len(_MGPU_TRACKED_MODELPATCHERS)})")
+
+def log_tracked_modelpatchers_status(tag="checkpoint"):
+    """Log count and estimated CPU RAM for tracked ModelPatchers"""
+    alive_count = len(_MGPU_TRACKED_MODELPATCHERS)
+    total_cpu_memory_mb = 0.0
+    
+    for patcher in list(_MGPU_TRACKED_MODELPATCHERS):
+        if hasattr(patcher, "model") and patcher.model is not None:
+            for param in patcher.model.parameters():
+                if getattr(param, "device", torch.device("cpu")).type == "cpu":
+                    total_cpu_memory_mb += (param.nelement() * param.element_size()) / (1024.0 * 1024.0)
+    
+    logger.warning(f"[MultiGPU_Lifecycle] [{tag}] Tracked ModelPatchers={alive_count}, approx CPU RAM={total_cpu_memory_mb:.2f} MB")
+
+def analyze_cpu_memory_leaks():
+    """Diagnostic: scan referrers of tracked ModelPatchers when memory is high"""
+    vm = psutil.virtual_memory()
+    patchers = list(_MGPU_TRACKED_MODELPATCHERS)
+    
+    if len(patchers) <= 5 and vm.percent <= 80.0:
+        logger.debug(f"[MultiGPU_Leak_Analyzer] Skipping analysis. Normal conditions: patchers={len(patchers)}, memory={vm.percent:.1f}%")
+        return
+        
+    logger.warning(f"[MultiGPU_Leak_Analyzer] High pressure detected: patchers={len(patchers)}, cpu_mem={vm.percent:.1f}%. Analyzing referrers.")
+    
+    for i, patcher in enumerate(patchers[:5]):
+        referrers = gc.get_referrers(patcher)
+        logger.warning(f"[MultiGPU_Leak_Analyzer] Patcher #{i} id={id(patcher)} referrers={len(referrers)}")
+        
+        for j, ref in enumerate(referrers[:10]):
+            rtype = type(ref).__name__
+            rmod = getattr(type(ref), "__module__", "unknown")
+            if isinstance(ref, dict):
+                logger.warning(f"  Ref {j}: dict(len={len(ref)}) mod={rmod}")
+            elif isinstance(ref, list):
+                logger.warning(f"  Ref {j}: list(len={len(ref)}) mod={rmod}")
+            else:
+                logger.warning(f"  Ref {j}: {rtype} mod={rmod}")
+
+# ==========================================================================================
+# Memory Management and Cleanup
+# ==========================================================================================
+
+CPU_MEMORY_THRESHOLD_PERCENT = 85.0
+CPU_RESET_HYSTERESIS_PERCENT = 5.0
+_last_cpu_usage_at_reset = 0.0
+
+def try_malloc_trim():
+    """Return freed heap memory to OS (Linux/glibc)"""
+    if platform.system() != "Linux":
+        return
+        
+    libc = ctypes.CDLL("libc.so.6")
+    if not hasattr(libc, "malloc_trim"):
+        return
+        
+    logger.info("[MultiGPU_Memory_Management] malloc_trim(0) begin")
+    multigpu_memory_log("mem_mgmt", "pre-malloc-trim")
+    
+    result = libc.malloc_trim(0)
+    
+    multigpu_memory_log("mem_mgmt", "post-malloc-trim")
+    if result == 1:
+        logger.info("[MultiGPU_Memory_Management] malloc_trim(0) released memory")
+    else:
+        logger.debug("[MultiGPU_Memory_Management] malloc_trim(0) no release")
+
+def trigger_executor_cache_reset(reason="policy", force=False):
+    """Trigger PromptExecutor.reset() by setting 'free_memory' flag"""
+    global _last_cpu_usage_at_reset
+
+    prompt_server = server.PromptServer.instance
+    if prompt_server is None:
+        logger.debug("[MultiGPU_Memory_Management] PromptServer not initialized")
+        return
+
+    if prompt_server.prompt_queue.currently_running and not force:
+        logger.debug(f"[MultiGPU_Memory_Management] Skipping reset during execution (reason: {reason})")
+        return
+
+    multigpu_memory_log("executor_reset", f"pre-trigger ({reason})")
+    logger.info(f"[MultiGPU_Memory_Management] Triggering PromptExecutor cache reset. Reason: {reason}")
+
+    analyze_cpu_memory_leaks()
+    prune_distorch_stores()
+    clear_memory_snapshot_history()
+
+    prompt_server.prompt_queue.set_flag("free_memory", True)
+    logger.debug("[MultiGPU_Memory_Management] 'free_memory' flag set")
+
+    vm = psutil.virtual_memory()
+    _last_cpu_usage_at_reset = vm.percent
+
+    try_malloc_trim()
+    multigpu_memory_log("executor_reset", f"post-trigger ({reason})")
+
+def check_cpu_memory_threshold(threshold_percent=CPU_MEMORY_THRESHOLD_PERCENT):
+    """Check CPU memory and trigger reset if threshold exceeded"""
+    if server.PromptServer.instance is None:
+        return
+
+    if server.PromptServer.instance.prompt_queue.currently_running:
+        return
+
+    vm = psutil.virtual_memory()
+    current_usage = vm.percent
+
+    if current_usage > threshold_percent:
+        if current_usage > (_last_cpu_usage_at_reset + CPU_RESET_HYSTERESIS_PERCENT):
+            logger.warning(f"[MultiGPU_Memory_Monitor] CPU usage ({current_usage:.1f}%) exceeds threshold ({threshold_percent:.1f}%)")
+            multigpu_memory_log("cpu_monitor", f"trigger:{current_usage:.1f}pct")
+            trigger_executor_cache_reset(reason="cpu_threshold_exceeded", force=False)
+        else:
+            logger.debug(f"[MultiGPU_Memory_Monitor] CPU usage high ({current_usage:.1f}%) but within hysteresis")
+            multigpu_memory_log("cpu_monitor", f"skip_hysteresis:{current_usage:.1f}pct")
+
+def force_full_system_cleanup(reason="manual", force=True):
+    """
+    Mirror ComfyUI-Manager 'Free model and node cache' by setting both flags:
+    unload_models=True and free_memory=True
+    """
+    vm = psutil.virtual_memory()
+    pre_cpu = vm.used
+    pre_models = len(mm.current_loaded_models)
+
+    multigpu_memory_log("full_cleanup", f"start:{reason}")
+    logger.mgpu_mm_log(f"[ManagerMatch] Requesting cleanup (reason={reason}) | pre_models={pre_models}, cpu_used_gib={pre_cpu/(1024**3):.2f}")
+
+    if server.PromptServer.instance is not None:
+        pq = server.PromptServer.instance.prompt_queue
+        if (not pq.currently_running) or force:
+            pq.set_flag("unload_models", True)
+            pq.set_flag("free_memory", True)
+            logger.mgpu_mm_log("[ManagerMatch] Flags set: unload_models=True, free_memory=True")
+        else:
+            logger.mgpu_mm_log("[ManagerMatch] Skipped - execution active and force=False")
+
+    vm = psutil.virtual_memory()
+    post_cpu = vm.used
+    post_models = len(mm.current_loaded_models)
+    delta_cpu_mb = (post_cpu - pre_cpu) / (1024**2)
+
+    multigpu_memory_log("full_cleanup", f"requested:{reason}")
+    summary = f"[ManagerMatch] Cleanup requested (reason={reason}) | models {pre_models}->{post_models}, cpu_delta_mb={delta_cpu_mb:.2f}"
+    logger.mgpu_mm_log(summary)
+    return summary
